fix: revert old-style migrations
