+++ peewee_migrate/template.txt
@@ -24,8 +24,6 @@ import datetime as dt
 
 import peewee as pw
 
-from peewee_migrate import migrations
-
 try:
     import playhouse.postgres_ext as pw_pext
 except ImportError:
@@ -34,15 +32,15 @@ except ImportError:
 SQL = pw.SQL
 
 
-class Migration(migrations.Migration):
-    atomic = True
+# Run the migration inside a single transaction
+__ATOMIC = True
 
-    @staticmethod
-    def migrate(migrator, database, fake=False):
-        """Write your migrations here."""
-    {migrate}
 
-    @staticmethod
-    def rollback(migrator, database, fake=False):
-        """Write your rollback migrations here."""
-    {rollback}
+def migrate(migrator, database, fake=False):
+    """Write your migrations here."""
+{migrate}
+
+
+def rollback(migrator, database, fake=False):
+    """Write your rollback migrations here."""
+{rollback}
